temp commit. booleans replaced with integers.

+++ hconfigure/header/ConfigHelpers.hpp
@@ -31,8 +31,8 @@ template <typename T> void writeIncDirsAtConfigTime(const vector<T> &include, Va
     {
         const InclNode &inclNode = getNode(elem);
         value.PushBack(inclNode.node->getValue(), rapidJsonAllocator);
-        value.PushBack(inclNode.isStandard, rapidJsonAllocator);
-        value.PushBack(inclNode.ignoreHeaderDeps, rapidJsonAllocator);
+        value.PushBack(static_cast<uint64_t>(inclNode.isStandard), rapidJsonAllocator);
+        value.PushBack(static_cast<uint64_t>(inclNode.ignoreHeaderDeps), rapidJsonAllocator);
     }
 }
